refactor(deploy): move state resolution into status.ts, rename dnsComplete
